+++ src/onchange_all_checkbox_.d.ts
@@ -1,6 +1,0 @@
-import { type WritableAtom_ } from '@ctx-core/nanostores'
-import type { checkbox_tree_default_Event_T } from './_types'
-export declare function onchange_all_checkbox_<Evt = checkbox_tree_default_Event_T>(
-	store$:WritableAtom_<Record<string, boolean>>, fn:(item_h_selected:Record<string, boolean>, checked:boolean)=>void
-):(event:Evt)=>void;
-export { onchange_all_checkbox_ as _onchange_all_checkbox, }