+++ src/onchange_checkbox_.d.ts
@@ -1,6 +1,0 @@
-import { type WritableAtom_ } from '@ctx-core/nanostores'
-import { checkbox_tree_default_Event_T } from './_types'
-export declare function onchange_checkbox_<Evt = checkbox_tree_default_Event_T>(
-	store$:WritableAtom_<Record<string, boolean>>
-):(event:Evt, item:string)=>void;
-export { onchange_checkbox_ as _onchange_checkbox, }